merging

+++ src/Controller/ProfilesController.php
@@ -2,6 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\ORM\TableRegistry;
 
 /**
  * Profiles Controller
@@ -116,16 +117,47 @@ class ProfilesController extends AppController
             }
             $this->Flash->error(__('The profile could not be saved. Please, try again.'));
         }
+        $status =array(""=>"Select","1"=>"Approved","2"=>"Not Approved");
         $users = $this->Profiles->Users->find('list', ['limit' => 200]);
-        $this->set(compact('profile', 'users'));
+        $this->set(compact('profile', 'users','status'));
     }
 
     public function addmy()
     {
         $profile = $this->Profiles->newEntity();
         if ($this->request->is('post')) {
+            $dir = \Cake\Core\Configure::read('App.wwwRoot');
+            $upLoadsDirectory = $dir.'/img/profile';
+
+            if (!file_exists($upLoadsDirectory)) {
+                mkdir($upLoadsDirectory, 0777, true);
+            }
+
+            //for picture
+            $fileParams = $this->request->data['images'];
+            $info = pathinfo($fileParams['name']);
+            $pathPicture = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+            if (!move_uploaded_file($this->request->data['images']['tmp_name'], $upLoadsDirectory.'/' . $pathPicture)) {
+                var_dump('Cant move picture ');
+                die;
+            }
+            unset($this->request->data['images']);
+            //end
+            //start signature
+            $fileParams = $this->request->data['sgn'];
+            $info = pathinfo($fileParams['name']);
+            $pathSign = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+            if (!move_uploaded_file($this->request->data['sgn']['tmp_name'], $upLoadsDirectory.'/' . $pathSign)) {
+                var_dump('Cant move signature ');
+                die;
+            }
+            unset($this->request->data['sgn']);
+            //end
             $profile = $this->Profiles->patchEntity($profile, $this->request->getData());
             $profile->user_id = $this->_userId();
+            $profile->picture = $pathPicture;
+            $profile->sgn = $pathSign;
+            $profile->status = 2;
             if ($this->Profiles->save($profile)) {
                 $this->Flash->success(__('The profile has been saved.'));
 
@@ -133,7 +165,8 @@ class ProfilesController extends AppController
             }
             $this->Flash->error(__('The profile could not be saved. Please, try again.'));
         }
-        $users = $this->Profiles->Users->find('list', ['limit' => 200]);
+        $users = TableRegistry::get('Users')->find('all')->where(['id'=>$this->_userId()])->first();
+
         $this->set(compact('profile', 'users'));
     }
     /**
@@ -146,10 +179,53 @@ class ProfilesController extends AppController
     public function edit($id = null)
     {
         $profile = $this->Profiles->get($id, [
-            'contain' => []
+            'contain' => ['Users']
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
+            // copy paste
+            $dir = \Cake\Core\Configure::read('App.wwwRoot');
+            $upLoadsDirectory = $dir.'/img/profile';
+
+            if (!file_exists($upLoadsDirectory)) {
+                mkdir($upLoadsDirectory, 0777, true);
+            }
+            //end :)
+            //for picture
+            $fileParams = $this->request->data['images'];
+            $info = pathinfo($fileParams['name']);
+            $pathPicture = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+
+            if(!empty($pathPicture) && !empty($this->request->data['images']['name'])){
+                $imageTrue ='ok';
+                if (!move_uploaded_file($this->request->data['images']['tmp_name'], $upLoadsDirectory.'/' . $pathPicture)) {
+                    var_dump('Cant move picture ');
+                    die;
+                }
+            }
+
+            unset($this->request->data['images']);
+            //end
+            //start signature
+            $fileParams = $this->request->data['sgn'];
+            $info = pathinfo($fileParams['name']);
+            $pathSign = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+
+            if(!empty($pathSign) && !empty($this->request->data['sgn']['name'])){
+                $signTrue ='ok';
+            if (!move_uploaded_file($this->request->data['sgn']['tmp_name'], $upLoadsDirectory.'/' . $pathSign)) {
+                var_dump('Cant move signature ');
+                die;
+            }
+            }
+            unset($this->request->data['sgn']);
+            //end
+
             $profile = $this->Profiles->patchEntity($profile, $this->request->getData());
+
+            if(!empty($pathPicture) && !empty($imageTrue)){$profile->picture = $pathPicture;}
+
+            if(!empty($pathSign) && !empty($signTrue)){$profile->sgn = $pathSign;}
+
             if ($this->Profiles->save($profile)) {
                 $this->Flash->success(__('The profile has been saved.'));
 
@@ -157,17 +233,60 @@ class ProfilesController extends AppController
             }
             $this->Flash->error(__('The profile could not be saved. Please, try again.'));
         }
+        $status =array(""=>"Select","1"=>"Approved","2"=>"Not Approved");
         $users = $this->Profiles->Users->find('list', ['limit' => 200]);
-        $this->set(compact('profile', 'users'));
+        $this->set(compact('profile', 'users','status'));
     }
 
     public function editmy($id = null)
     {
         $profile = $this->Profiles->get($id, [
-            'contain' => []
+            'contain' => ['Users']
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
+            // copy paste
+            $dir = \Cake\Core\Configure::read('App.wwwRoot');
+            $upLoadsDirectory = $dir.'/img/profile';
+
+            if (!file_exists($upLoadsDirectory)) {
+                mkdir($upLoadsDirectory, 0777, true);
+            }
+            //end :)
+            //for picture
+            $fileParams = $this->request->data['images'];
+            $info = pathinfo($fileParams['name']);
+            $pathPicture = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+
+            if(!empty($pathPicture) && !empty($this->request->data['images']['name'])){
+                $imageTrue ='ok';
+                if (!move_uploaded_file($this->request->data['images']['tmp_name'], $upLoadsDirectory.'/' . $pathPicture)) {
+                    var_dump('Cant move picture ');
+                    die;
+                }
+            }
+
+            unset($this->request->data['images']);
+            //end
+            //start signature
+            $fileParams = $this->request->data['sgn'];
+            $info = pathinfo($fileParams['name']);
+            $pathSign = md5($fileParams['name']) . '-' . uniqid() . '.' . $info['extension'];
+
+            if(!empty($pathSign) && !empty($this->request->data['sgn']['name'])){
+                $signTrue ='ok';
+                if (!move_uploaded_file($this->request->data['sgn']['tmp_name'], $upLoadsDirectory.'/' . $pathSign)) {
+                    var_dump('Cant move signature ');
+                    die;
+                }
+            }
+            unset($this->request->data['sgn']);
+            //end
+
             $profile = $this->Profiles->patchEntity($profile, $this->request->getData());
+
+            if(!empty($pathPicture) && !empty($imageTrue)){$profile->picture = $pathPicture;}
+
+            if(!empty($pathSign) && !empty($signTrue)){$profile->sgn = $pathSign;}
             $profile->user_id = $this->_userId();
             if ($this->Profiles->save($profile)) {
                 $this->Flash->success(__('The profile has been saved.'));
@@ -191,12 +310,17 @@ class ProfilesController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $profile = $this->Profiles->get($id);
-        if ($this->Profiles->delete($profile)) {
-            $this->Flash->success(__('The profile has been deleted.'));
-        } else {
-            $this->Flash->error(__('The profile could not be deleted. Please, try again.'));
+        //if $profile not found, redirect back to list view
+        if (empty($profile)) {
+            $this->Flash->error(__('User not found'));
+            return $this->redirect(['action' => 'index']);
         }
 
+        if ($this->Profiles->delete($profile)) {
+            return $this->response->withType("application/json")->withStringBody(json_encode(array('status' => 'deleted'))); die;
+        } else {
+            return $this->response->withType("application/json")->withStringBody(json_encode(array('status' => 'error'))); die;
+        }
         return $this->redirect(['action' => 'index']);
     }
 }
